First Responder
I’ve added a property to choose if the control should handle first
responder or not. I’ve done this because I was losing the focus on a
text view while using the control.
I don’t know why the control do this, but instead to remove everything
related to firstResponder, I use a flag.